requirements.txt
line-bot-sdk
gunicorn
flask
beautifulsoup4
requests